Unidad de medidas, correciones del botón guardar y cancelar, decimales en productos y iva

+++ app/Http/Controllers/ProductosController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\Redirect;
 use Illuminate\Http\Request;
 use App\Models\Categorias;
 use App\Models\Ivas;
+use App\Models\Medidas;
 
 class ProductosController extends Controller
 {
@@ -16,11 +17,13 @@ class ProductosController extends Controller
     public function index()
     {
         $categorias = Categorias::all();
+        $medidas = Medidas::all();
         $ivas = Ivas::all();
 
         $data = [
             'ivas' => $ivas,
             'categorias' => $categorias,
+            'medidas' => $medidas,
        ];
 
         return view('productos.create')->with($data);     
@@ -28,20 +31,24 @@ class ProductosController extends Controller
 
     public function create(ProductoRequest $request)
     {
+      
         $productos = new Productos();
         $productos->name = $request->name;
-        $productos->precio_sin_iva = $request->precio_sin_iva;
-        $productos->costo_unitario = $request->costo_unitario;
+
         $productos->contenido_neto = $request->contenido_neto;
         $productos->unidad = $request->unidad;
-        $productos->peso = $request->peso;
         $productos->altura = $request->altura;
         $productos->ancho = $request->ancho;
-        $productos->longitud = $request->longitud;
         $productos->description = $request->description;
         $productos->upc = $request->upc;
         $productos->id_categoria = $request->id_categoria;
+        $productos->id_medida = $request->medida;
         $productos->exento = $request->exento;
+
+
+
+        $productos->precio_sin_iva = str_replace(",",".",$request->precio_sin_iva);
+        $productos->costo_unitario = str_replace(",",".",$request->costo_unitario);
 
         if($request->hasFile("imagen_url")){
 
@@ -63,11 +70,22 @@ class ProductosController extends Controller
        
         $categorias = Categorias::all();
         $ivas = Ivas::all();
+        $medidas = Medidas::all();
         $producto = Productos::find($id);
+
+        $producto->precio_sin_iva = number_format($producto->precio_sin_iva, 2);
+        $producto->precio_sin_iva = str_replace(",","",$producto->precio_sin_iva);
+        $producto->precio_sin_iva = str_replace(".",",",$producto->precio_sin_iva);
+
+
+        $producto->costo_unitario = number_format($producto->costo_unitario, 2);
+        $producto->costo_unitario = str_replace(",","",$producto->costo_unitario);
+        $producto->costo_unitario = str_replace(".",",",$producto->costo_unitario);
 
         $data = [
             'ivas' => $ivas,
             'categorias' => $categorias,
+            'medidas' => $medidas,
             'producto' => $producto
        ];
 
@@ -81,27 +99,32 @@ class ProductosController extends Controller
         $producto = Productos::find($request->id);
         $producto->name = $request->name;
 
-        $producto->precio_sin_iva = $request->precio_sin_iva;
-        $producto->costo_unitario = $request->costo_unitario;
         $producto->contenido_neto = $request->contenido_neto;
         $producto->unidad = $request->unidad;
-        $producto->peso = $request->peso;
         $producto->altura = $request->altura;
         $producto->ancho = $request->ancho;
-        $producto->longitud = $request->longitud;
         $producto->description = $request->description;
         $producto->upc = $request->upc;
         $producto->id_categoria = $request->id_categoria;
         $producto->exento = $request->exento;
+        $producto->id_medida = $request->medida;
+
+
+        $producto->precio_sin_iva = str_replace(",",".",$request->precio_sin_iva);
+        $producto->costo_unitario = str_replace(",",".",$request->costo_unitario);
 
         if($request->hasFile("imagen_url")){
             /*****eliminar la img anterior */ 
+
+            if($producto->imagen!= null){
             unlink(public_path('app/archivos/productos/'.$producto->imagen));
+            }
             $imagen = $request->file("imagen_url");
             $nombreimagen =  $imagen->getClientOriginalName();
             $ruta = public_path("app/archivos/productos/");
             $imagen->move($ruta,$nombreimagen);
-            $producto->imagen = $nombreimagen;
+            $producto->imagen_url = $nombreimagen;
+            
         }
             $producto->save();
             return Redirect::route('productos');
